fix: 링크 공유 로직 custom hook으로 추상화
